Add identity rename mechanism
Why did we feel justified in adding this complexity?
Currently our users can use any username, but we want to force them to use
email addresses instead. We believe that forcing them to create new
accounts is to disruptive, so we needed to create a mechanism that allows
them to rename their identities.