TI-3497: Fix for ldap user update loop (#42)
* TI-3497: Fix for ldap user update loop
A ldap user with a space at the end of his email caused the system to update him
during every ldap merge as the space at the end of the email would mysteriously
dissapear after the user structs address was taken. We trim the spaces of the
emails now by default. Add exstra loging for ldap merge user update.

* TI-3497: Adds space trimming to all ldap attribs
Adds warning comments about possible compiler bug that removes the last space
from the email attrib in the user struct

* Fix spelling mistake

* Changed json fields for AuthUser to camelCase