Add Azure Active Directory synchronization
The MSAAD (Microsoft Azure Active Directory) sync is a new concept that
is very similar in spirit to the LDAP sync that already exists.
We read from an MSAAD and synchronize the users from there, into the
Authaus database.

Also, changed to OAuth:
1. Added continual validation of the OAuth tokens, so that if the OAuth
session is invalidated, then the associated Authaus session is also
invalidated.
2. Added cleanup systems for clearing out stale OAuth data.
3. Made the creation of new users from an OAuth login optional, and off
by default.
4. Fixed a bug in the refresh of the OAuth token