Changes to the sql_db to remove indentation
Navigating all the branches in some of the old format is quite an
interesting (read: confusion) experience. This commit seeks to amend
that by inverting some of the conditional to allow for quick escaping
and ease of visualization of where the prospective errors may lie

This commit also converts some of the Infof calls to Errorf calls,
because the noise is too great

Next steps: Deal with how VerboseLogging is handled in the service so
that they are logged as Trace calls, rather than Info calls.