Allow configuration for number of historic passwords
This is used in combination with the feature that prevents the user
from re-using old passwords.

Previously, the number of old passwords was hardcoded to 15. The default
remains 15, but the precise number of old passwords is now configurable.

Also added a bunch of log messages during startup, to help diagnose
if you've made a configuration error.