Introduce pseudo UserId values for LDAP/MSAAD/OAuth user merge systems
This is a fix for ticket INF-1369.

The audit trail needs more information about who/what is modifying
a user, which is why we're introducing 3 new pseudo UserId values:

 0: Administrator (this was already present)
-2: LDAP Merge
-3: OAuth Merge
-4: MSAAD Merge

This commit also fixes a host of broken unit tests, so that all tests
now at least pass with the Postgres backend, and dummy backend.

NOTE: The following commit contains the *actual* fix for the bug
mentioned. This commit lays the groundwork for a proper fix.