Fix incorrect "ModifiedBy" value resulting from LDAP merge
This is a fix for ticket INF-1369

The real bug here is that user.Modified was altered, but user.ModifiedBy
was inherited from the previous version of the record. So when a user
popped open the auth UI, it would say that this record was modified
by whoever previously modified the record, but the date would be a modern
date (ie whenever the LDAP merge process last made any change).

The customer complained because the person who supposedly made recent
changes to various users, had not even logged in during that time period.