Add authuserstore db triggers. #59
fix: Add authuserstore db triggers.

Add triggers to ensure that created, createdby and modified for the authuserstore table gets set correctly.

created - this gets set to NOW() when a record gets inserted. Once the record has been inserted this value should not change and the update trigger ensures this.
createdby - if this value is null when a record gets inserted an exception will be raised. Once the record has been inserted this value should not change and the update trigger ensures this.
modified - this value should always be set to NOW() whenever a record gets inserted or updated.
The naming convention used for the functions and triggers is as follows:
{database name}{schema name}{table name}_{type of trigger}

type of trigger:

first character denoted when the action happens. 'b' for before, 'a' for after.
second character denoted the type of action. 'i' for insert, 'u' for update.
the last character just indicates that this is for a trigger and 't' is used.
Jira Ticket: ASG-2210